# docs/api/payments-openapi.yaml
openapi: 3.0.1
info:
  title: Payments Service API
  description: Saga payment processing (consumes order.placed) and payment read access.
  version: v1
servers:
- url: http://localhost:8084
  description: Generated server url
tags:
- name: Payments
  description: Read access to saga-driven payments
paths:
  /api/v1/payments:
    get:
      tags:
      - Payments
      summary: "List all payments [ADMIN]"
      operationId: all
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentResponse"
      security:
      - bearerAuth: []
  /api/v1/payments/{id}:
    get:
      tags:
      - Payments
      summary: Get a payment by id (owner or ADMIN)
      operationId: byId
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaymentResponse"
      security:
      - bearerAuth: []
  /api/v1/payments/order/{orderId}:
    get:
      tags:
      - Payments
      summary: Get the payment for an order (owner or ADMIN)
      operationId: byOrderId
      parameters:
      - name: orderId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaymentResponse"
      security:
      - bearerAuth: []
components:
  schemas:
    PaymentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        orderId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        amount:
          type: number
        status:
          type: string
        method:
          type: string
        last4:
          type: string
        failureReason:
          type: string
        createdAt:
          type: string
          format: date-time
  securitySchemes:
    bearerAuth:
      type: http
      description: Access token from the Users service login
      scheme: bearer
      bearerFormat: JWT